+++ services/executionlayer/src/main/java/tech/pegasys/teku/services/executionlayer/ExecutionLayerConfiguration.java
@@ -42,6 +42,7 @@ public class ExecutionLayerConfiguration {
   private final Spec spec;
   private final Optional<String> engineEndpoint;
   private final Optional<String> engineJwtSecretFile;
+  private final Optional<String> engineJwtClaimId;
   private final Optional<String> builderEndpoint;
   private final boolean isBuilderCircuitBreakerEnabled;
   private final int builderCircuitBreakerWindow;
@@ -56,6 +57,7 @@ public class ExecutionLayerConfiguration {
       final Spec spec,
       final Optional<String> engineEndpoint,
       final Optional<String> engineJwtSecretFile,
+      final Optional<String> engineJwtClaimId,
       final Optional<String> builderEndpoint,
       final boolean isBuilderCircuitBreakerEnabled,
       final int builderCircuitBreakerWindow,
@@ -68,6 +70,7 @@ public class ExecutionLayerConfiguration {
     this.spec = spec;
     this.engineEndpoint = engineEndpoint;
     this.engineJwtSecretFile = engineJwtSecretFile;
+    this.engineJwtClaimId = engineJwtClaimId;
     this.builderEndpoint = builderEndpoint;
     this.isBuilderCircuitBreakerEnabled = isBuilderCircuitBreakerEnabled;
     this.builderCircuitBreakerWindow = builderCircuitBreakerWindow;
@@ -101,6 +104,10 @@ public class ExecutionLayerConfiguration {
 
   public Optional<String> getEngineJwtSecretFile() {
     return engineJwtSecretFile;
+  }
+
+  public Optional<String> getEngineJwtClaimId() {
+    return engineJwtClaimId;
   }
 
   public Optional<String> getBuilderEndpoint() {
@@ -143,6 +150,7 @@ public class ExecutionLayerConfiguration {
     private Spec spec;
     private Optional<String> engineEndpoint = Optional.empty();
     private Optional<String> engineJwtSecretFile = Optional.empty();
+    private Optional<String> engineJwtClaimId = Optional.empty();
     private Optional<String> builderEndpoint = Optional.empty();
     private boolean isBuilderCircuitBreakerEnabled = DEFAULT_BUILDER_CIRCUIT_BREAKER_ENABLED;
     private int builderCircuitBreakerWindow = DEFAULT_BUILDER_CIRCUIT_BREAKER_WINDOW;
@@ -182,6 +190,7 @@ public class ExecutionLayerConfiguration {
           spec,
           engineEndpoint,
           engineJwtSecretFile,
+          engineJwtClaimId,
           builderEndpoint,
           isBuilderCircuitBreakerEnabled,
           builderCircuitBreakerWindow,
@@ -205,6 +214,11 @@ public class ExecutionLayerConfiguration {
 
     public Builder engineJwtSecretFile(final String jwtSecretFile) {
       this.engineJwtSecretFile = Optional.ofNullable(jwtSecretFile).filter(StringUtils::isNotBlank);
+      return this;
+    }
+
+    public Builder engineJwtClaimId(final String jwtClaimId) {
+      this.engineJwtClaimId = Optional.ofNullable(jwtClaimId).filter(StringUtils::isNotBlank);
       return this;
     }